requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
httpx
tenacity
loguru
